Finalize refactoring
