+++ internal/indexer/indexer_test.go
@@ -1,7 +1,6 @@
 package indexer
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"github.com/creekorful/trandoshan/api"
@@ -149,14 +148,6 @@ func TestAddResource(t *testing.T) {
 		Description: "the description",
 		Headers:     map[string]string{"Content-Type": "application/html", "Server": "Traefik"},
 	}
-	bodyBytes, err := json.Marshal(body)
-	if err != nil {
-		t.FailNow()
-	}
-
-	// The requests
-	req := httptest.NewRequest(http.MethodPost, "/v1/resources", bytes.NewReader(bodyBytes))
-	rec := httptest.NewRecorder()
 
 	dbMock := database_mock.NewMockDatabase(mockCtrl)
 	configClientMock := client_mock.NewMockClient(mockCtrl)
@@ -192,17 +183,8 @@ func TestAddResource(t *testing.T) {
 	configClientMock.EXPECT().GetForbiddenHostnames().Return([]client.ForbiddenHostname{}, nil)
 
 	s := State{db: dbMock, configClient: configClientMock, pub: pubMock}
-
-	s.addResource(rec, req)
-	if rec.Code != http.StatusOK {
-		t.FailNow()
-	}
-	if rec.Header().Get("Content-Type") != "application/json" {
-		t.Fail()
-	}
-
-	var res api.ResourceDto
-	if err := json.NewDecoder(rec.Body).Decode(&res); err != nil {
+	res, err := s.addResource(body)
+	if err != nil {
 		t.FailNow()
 	}
 
@@ -245,14 +227,6 @@ func TestAddResourceDuplicateNotAllowed(t *testing.T) {
 		Description: "the description",
 		Headers:     map[string]string{"Content-Type": "application/html", "Server": "Traefik"},
 	}
-	bodyBytes, err := json.Marshal(body)
-	if err != nil {
-		t.FailNow()
-	}
-
-	// The requests
-	req := httptest.NewRequest(http.MethodPost, "/v1/resources", bytes.NewReader(bodyBytes))
-	rec := httptest.NewRecorder()
 
 	dbMock := database_mock.NewMockDatabase(mockCtrl)
 	configClientMock := client_mock.NewMockClient(mockCtrl)
@@ -268,8 +242,7 @@ func TestAddResourceDuplicateNotAllowed(t *testing.T) {
 
 	s := State{db: dbMock, configClient: configClientMock}
 
-	s.addResource(rec, req)
-	if rec.Code != http.StatusOK {
+	if _, err := s.addResource(body); err != errAlreadyIndexed {
 		t.FailNow()
 	}
 }
@@ -287,14 +260,6 @@ func TestAddResourceTooYoung(t *testing.T) {
 		Description: "the description",
 		Headers:     map[string]string{"Content-Type": "application/html", "Server": "Traefik"},
 	}
-	bodyBytes, err := json.Marshal(body)
-	if err != nil {
-		t.FailNow()
-	}
-
-	// The requests
-	req := httptest.NewRequest(http.MethodPost, "/v1/resources", bytes.NewReader(bodyBytes))
-	rec := httptest.NewRecorder()
 
 	dbMock := database_mock.NewMockDatabase(mockCtrl)
 	configClientMock := client_mock.NewMockClient(mockCtrl)
@@ -311,8 +276,7 @@ func TestAddResourceTooYoung(t *testing.T) {
 
 	s := State{db: dbMock, configClient: configClientMock}
 
-	s.addResource(rec, req)
-	if rec.Code != http.StatusOK {
+	if _, err := s.addResource(body); err != errAlreadyIndexed {
 		t.FailNow()
 	}
 }
@@ -330,14 +294,6 @@ func TestAddResourceForbiddenHostname(t *testing.T) {
 		Description: "the description",
 		Headers:     map[string]string{"Content-Type": "application/html", "Server": "Traefik"},
 	}
-	bodyBytes, err := json.Marshal(body)
-	if err != nil {
-		t.FailNow()
-	}
-
-	// The requests
-	req := httptest.NewRequest(http.MethodPost, "/v1/resources", bytes.NewReader(bodyBytes))
-	rec := httptest.NewRecorder()
 
 	configClientMock := client_mock.NewMockClient(mockCtrl)
 
@@ -345,8 +301,7 @@ func TestAddResourceForbiddenHostname(t *testing.T) {
 
 	s := State{configClient: configClientMock}
 
-	s.addResource(rec, req)
-	if rec.Code != http.StatusOK {
+	if _, err := s.addResource(body); err != errHostnameNotAllowed {
 		t.FailNow()
 	}
 }
@@ -396,6 +351,186 @@ func TestSearchResources(t *testing.T) {
 	}
 	if len(resources) != 2 {
 		t.Errorf("got %d resources want 2", len(resources))
+	}
+}
+
+func TestExtractResource(t *testing.T) {
+	body := `
+<title>Creekorful Inc</title>
+
+This is sparta
+
+<a href="https://google.com/test?test=test#12">
+
+<meta name="Description" content="Zhello world">
+<meta property="og:url" content="https://example.org">
+`
+
+	msg := event.NewResourceEvent{
+		URL:  "https://example.org/300",
+		Body: body,
+	}
+
+	resDto, urls, err := extractResource(msg)
+	if err != nil {
+		t.FailNow()
+	}
+
+	if resDto.URL != "https://example.org/300" {
+		t.Fail()
+	}
+	if resDto.Title != "Creekorful Inc" {
+		t.Fail()
+	}
+	if resDto.Body != msg.Body {
+		t.Fail()
+	}
+
+	if len(urls) != 2 {
+		t.FailNow()
+	}
+	if urls[0] != "https://google.com/test?test=test" {
+		t.Fail()
+	}
+	if urls[1] != "https://example.org" {
+		t.Fail()
+	}
+
+	if resDto.Description != "Zhello world" {
+		t.Fail()
+	}
+
+	if resDto.Meta["description"] != "Zhello world" {
+		t.Fail()
+	}
+
+	if resDto.Meta["og:url"] != "https://example.org" {
+		t.Fail()
+	}
+}
+
+func TestNormalizeURL(t *testing.T) {
+	url, err := normalizeURL("https://this-is-sparta.de?url=url-query-param#fragment-23")
+	if err != nil {
+		t.FailNow()
+	}
+
+	if url != "https://this-is-sparta.de?url=url-query-param" {
+		t.Fail()
+	}
+}
+
+func TestHandleNewResourceEvent(t *testing.T) {
+	body := `
+<title>Creekorful Inc</title>
+
+This is sparta (hosted on https://example.org)
+
+<a href="https://google.com/test?test=test#12">
+
+<meta name="DescriptIon" content="Zhello world">
+<meta property="og:url" content="https://example.org">`
+
+	mockCtrl := gomock.NewController(t)
+	defer mockCtrl.Finish()
+
+	subscriberMock := event_mock.NewMockSubscriber(mockCtrl)
+	configClientMock := client_mock.NewMockClient(mockCtrl)
+	dbMock := database_mock.NewMockDatabase(mockCtrl)
+	pubMock := event_mock.NewMockPublisher(mockCtrl)
+
+	tn := time.Now()
+
+	msg := event.RawMessage{}
+	subscriberMock.EXPECT().
+		Read(&msg, &event.NewResourceEvent{}).
+		SetArg(1, event.NewResourceEvent{
+			URL:     "https://example.onion",
+			Body:    body,
+			Headers: map[string]string{"Server": "Traefik", "Content-Type": "application/html"},
+			Time:    tn,
+		}).Return(nil)
+
+	configClientMock.EXPECT().GetForbiddenHostnames().Return([]client.ForbiddenHostname{{Hostname: "example2.onion"}}, nil)
+	configClientMock.EXPECT().GetRefreshDelay().Return(client.RefreshDelay{Delay: -1}, nil)
+
+	dbMock.EXPECT().CountResources(&api.ResSearchParams{
+		URL:        "https://example.onion",
+		PageSize:   1,
+		PageNumber: 1,
+	}).Return(int64(0), nil)
+
+	// make sure we are creating the resource
+	dbMock.EXPECT().AddResource(database.ResourceIdx{
+		URL:         "https://example.onion",
+		Body:        body,
+		Title:       "Creekorful Inc",
+		Meta:        map[string]string{"description": "Zhello world", "og:url": "https://example.org"},
+		Description: "Zhello world",
+		Headers:     map[string]string{"server": "Traefik", "content-type": "application/html"},
+		Time:        tn,
+	}).Return(nil)
+
+	pubMock.EXPECT().PublishEvent(&event.NewIndexEvent{
+		URL:         "https://example.onion",
+		Body:        body,
+		Title:       "Creekorful Inc",
+		Meta:        map[string]string{"description": "Zhello world", "og:url": "https://example.org"},
+		Description: "Zhello world",
+		Headers:     map[string]string{"server": "Traefik", "content-type": "application/html"},
+		Time:        tn,
+	}).Return(nil)
+
+	// make sure we are pushing found URLs
+
+	// should be called only one time
+	subscriberMock.EXPECT().
+		PublishEvent(&event.FoundURLEvent{URL: "https://example.org"}).
+		Return(nil)
+	subscriberMock.EXPECT().
+		PublishEvent(&event.FoundURLEvent{URL: "https://google.com/test?test=test"}).
+		Return(nil)
+
+	s := State{db: dbMock, configClient: configClientMock, pub: pubMock}
+	if err := s.handleNewResourceEvent(subscriberMock, msg); err != nil {
+		t.FailNow()
+	}
+}
+
+func TestHandleMessageForbiddenHostname(t *testing.T) {
+	body := `
+<title>Creekorful Inc</title>
+
+This is sparta (hosted on https://example.org)
+
+<a href="https://google.com/test?test=test#12">
+
+<meta name="DescriptIon" content="Zhello world">
+<meta property="og:url" content="https://example.org">`
+
+	mockCtrl := gomock.NewController(t)
+	defer mockCtrl.Finish()
+
+	subscriberMock := event_mock.NewMockSubscriber(mockCtrl)
+	configClientMock := client_mock.NewMockClient(mockCtrl)
+
+	tn := time.Now()
+
+	msg := event.RawMessage{}
+	subscriberMock.EXPECT().
+		Read(&msg, &event.NewResourceEvent{}).
+		SetArg(1, event.NewResourceEvent{
+			URL:     "https://example.onion",
+			Body:    body,
+			Headers: map[string]string{"Server": "Traefik", "Content-Type": "application/html"},
+			Time:    tn,
+		}).Return(nil)
+
+	configClientMock.EXPECT().GetForbiddenHostnames().Return([]client.ForbiddenHostname{{Hostname: "example.onion"}}, nil)
+
+	s := State{configClient: configClientMock}
+	if err := s.handleNewResourceEvent(subscriberMock, msg); err != errHostnameNotAllowed {
+		t.FailNow()
 	}
 }
