+++ internal/scheduler/scheduler.go
@@ -3,7 +3,6 @@ package scheduler
 import (
 	"errors"
 	"fmt"
-	"github.com/creekorful/trandoshan/api"
 	configapi "github.com/creekorful/trandoshan/internal/configapi/client"
 	"github.com/creekorful/trandoshan/internal/constraint"
 	"github.com/creekorful/trandoshan/internal/event"
@@ -13,20 +12,17 @@ import (
 	"net/http"
 	"net/url"
 	"strings"
-	"time"
 )
 
 var (
 	errNotOnionHostname    = errors.New("hostname is not .onion")
 	errProtocolNotAllowed  = errors.New("protocol is not allowed")
 	errExtensionNotAllowed = errors.New("extension is not allowed")
-	errShouldNotSchedule   = errors.New("should not be scheduled")
 	errHostnameNotAllowed  = errors.New("hostname is not allowed")
 )
 
 // State represent the application state
 type State struct {
-	apiClient    api.API
 	configClient configapi.Client
 }
 
@@ -37,7 +33,7 @@ func (state *State) Name() string {
 
 // CommonFlags return process common flags
 func (state *State) CommonFlags() []string {
-	return []string{process.HubURIFlag, process.APIURIFlag, process.APITokenFlag, process.ConfigAPIURIFlag}
+	return []string{process.HubURIFlag, process.ConfigAPIURIFlag}
 }
 
 // CustomFlags return process custom flags
@@ -47,13 +43,7 @@ func (state *State) CustomFlags() []cli.Flag {
 
 // Initialize the process
 func (state *State) Initialize(provider process.Provider) error {
-	apiClient, err := provider.APIClient()
-	if err != nil {
-		return err
-	}
-	state.apiClient = apiClient
-
-	keys := []string{configapi.ForbiddenMimeTypesKey, configapi.ForbiddenHostnamesKey, configapi.RefreshDelayKey}
+	keys := []string{configapi.ForbiddenMimeTypesKey, configapi.ForbiddenHostnamesKey}
 	configClient, err := provider.ConfigClient(keys)
 	if err != nil {
 		return err
@@ -117,32 +107,6 @@ func (state *State) handleURLFoundEvent(subscriber event.Subscriber, msg event.R
 		return fmt.Errorf("%s %w", u, errHostnameNotAllowed)
 	}
 
-	// If we want to allow re-schedule of existing crawled resources we need to retrieve only resources
-	// that are newer than `now - refreshDelay`.
-	endDate := time.Time{}
-	if refreshDelay, err := state.configClient.GetRefreshDelay(); err == nil {
-		if refreshDelay.Delay != -1 {
-			endDate = time.Now().Add(-refreshDelay.Delay)
-		}
-	}
-
-	params := api.ResSearchParams{
-		URL:        u.String(),
-		EndDate:    endDate,
-		WithBody:   false,
-		PageSize:   1,
-		PageNumber: 1,
-	}
-	_, count, err := state.apiClient.SearchResources(&params)
-	if err != nil {
-		return fmt.Errorf("error while searching resource (%s): %s", u, err)
-	}
-
-	if count > 0 {
-		return fmt.Errorf("%s %w", u, errShouldNotSchedule)
-	}
-
-	// No matches: schedule!
 	log.Debug().Stringer("url", u).Msg("URL should be scheduled")
 
 	if err := subscriber.PublishEvent(&event.NewURLEvent{URL: evt.URL}); err != nil {
